RGW POSIX - Split FSEnt into it's own files
